Split state
